+++ src/main/java/dnf/Main.java
@@ -1,4 +1,4 @@
-package io.github.lunasaw;
+package dnf;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -9,6 +9,6 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Hello world!");
+        log.info("main::args = {}", args);
     }
 }
