+++ src/main/java/dnf/game/Address.java
@@ -1,0 +1,4 @@
+package dnf.game;
+
+public enum Address {
+}
