+++ src/main/java/dnf/game/Auto.java
@@ -1,0 +1,4 @@
+package dnf.game;
+
+public class Auto {
+}
