+++ src/main/java/dnf/game/Call.java
@@ -1,0 +1,4 @@
+package dnf.game;
+
+public class Call {
+}
