+++ src/main/java/dnf/game/Pack.java
@@ -1,0 +1,4 @@
+package dnf.game;
+
+public class Pack {
+}
