+++ src/main/java/dnf/game/Traverse.java
@@ -1,0 +1,4 @@
+package dnf.game;
+
+public class Traverse {
+}
