--- src/main/java/io/github/lunasaw/game/Address.java
@@ -1,4 +1,0 @@
-package io.github.lunasaw.game;
-
-public enum Address {
-}
