--- src/main/java/io/github/lunasaw/game/Pack.java
@@ -1,4 +1,0 @@
-package io.github.lunasaw.game;
-
-public class Pack {
-}
