--- src/main/java/io/github/lunasaw/game/Traverse.java
@@ -1,4 +1,0 @@
-package io.github.lunasaw.game;
-
-public class Traverse {
-}
